+++ djangojs/static/js/test/django.specs.js
@@ -6,26 +6,26 @@
 describe("Django.js", function(){
     var DJANGO_INFOS = window.DJANGO_INFOS;
 
-    it("should be defined", function(){
+    it("is defined", function(){
         expect(Django).toBeDefined();
     });
 
     describe('Context', function(){
-        it('should have a context attribute', function(){
+        it('have a context attribute', function(){
             expect(Django.context).toBeDefined();
         });
 
-        it('should store STATIC_URL constant', function(){
+        it('store STATIC_URL constant', function(){
             expect(Django.context.STATIC_URL).toBeDefined();
             expect(Django.context.STATIC_URL).toBe(DJANGO_INFOS.STATIC_URL);
         });
 
-        it('should store MEDIA_URL constant', function(){
+        it('store MEDIA_URL constant', function(){
             expect(Django.context.MEDIA_URL).toBeDefined();
             expect(Django.context.MEDIA_URL).toBe(DJANGO_INFOS.MEDIA_URL);
         });
 
-        it('should store available LANGUAGES', function(){
+        it('store available LANGUAGES', function(){
             expect(Django.context.LANGUAGES).toBeDefined();
             for (var code in DJANGO_INFOS.LANGUAGES) {
                 expect(Django.context.LANGUAGES[code]).toBeDefined();
@@ -33,27 +33,27 @@ describe("Django.js", function(){
             }
         });
 
-        it('should store LANGUAGE_CODE', function(){
+        it('store LANGUAGE_CODE', function(){
             expect(Django.context.LANGUAGE_CODE).toBeDefined();
             expect(Django.context.LANGUAGE_CODE).toBe(DJANGO_INFOS.LANGUAGE_CODE);
         });
 
-        it('should store LANGUAGE_BIDI', function(){
+        it('store LANGUAGE_BIDI', function(){
             expect(Django.context.LANGUAGE_BIDI).toBeDefined();
             expect(Django.context.LANGUAGE_BIDI).toBe(DJANGO_INFOS.LANGUAGE_BIDI === 'True');
         });
 
-        it('should store LANGUAGE_NAME', function(){
+        it('store LANGUAGE_NAME', function(){
             expect(Django.context.LANGUAGE_NAME).toBeDefined();
             expect(Django.context.LANGUAGE_NAME).toBe(DJANGO_INFOS.LANGUAGE_NAME);
         });
 
-        it('should store LANGUAGE_NAME_LOCAL', function(){
+        it('store LANGUAGE_NAME_LOCAL', function(){
             expect(Django.context.LANGUAGE_NAME_LOCAL).toBeDefined();
             expect(Django.context.LANGUAGE_NAME_LOCAL).toBe(DJANGO_INFOS.LANGUAGE_NAME_LOCAL);
         });
 
-        it('should store any custom context value', function(){
+        it('store any custom context value', function(){
             expect(Django.context.CUSTOM).toBeDefined();
             expect(Django.context.CUSTOM).toBe('CUSTOM_VALUE');
         });
@@ -61,7 +61,7 @@ describe("Django.js", function(){
 
     describe('Resolve reverse URLs', function(){
 
-        it('should throw if URL name does not exists', function(){
+        it('throw if URL name does not exists', function(){
             var notFoundUrl = function() {
                 Django.url('unknown');
             };
@@ -70,27 +70,27 @@ describe("Django.js", function(){
         });
 
         describe("from arguments", function(){
-            it("should resolve an URL without parameter", function(){
+            it("resolve an URL without parameter", function(){
                 expect(Django.url('test_form')).toBe('/test/form');
             });
 
-            it("should resolve an URL with an anonymous token ", function(){
+            it("resolve an URL with an anonymous token ", function(){
                 expect(Django.url('test_arg', 41)).toBe('/test/arg/41');
             });
 
-            it("should resolve an URL with many anonymous token", function(){
+            it("resolve an URL with many anonymous token", function(){
                 expect(Django.url('test_arg_multi', 41, 'test')).toBe('/test/arg/41/test');
             });
 
-            it("should resolve an URL with a named argument", function(){
+            it("resolve an URL with a named argument", function(){
                 expect(Django.url('test_named', 'test')).toBe('/test/named/test');
             });
 
-            it("should resolve an URL with many named token", function(){
+            it("resolve an URL with many named token", function(){
                 expect(Django.url('test_named_multi', 'value', 41)).toBe('/test/named/value/41');
             });
 
-            it('should throw if number of parameters mismatch', function(){
+            it('throw if number of parameters mismatch', function(){
                 var tooMany = function() {
                     Django.url('test_arg', 1, 2);
                 };
@@ -106,29 +106,29 @@ describe("Django.js", function(){
         });
 
         describe("from array", function(){
-            it("should resolve an URL without parameter", function(){
+            it("resolve an URL without parameter", function(){
                 expect(Django.url('test_form', [])).toBe('/test/form');
             });
 
-            it("should resolve an URL with an anonymous token ", function(){
+            it("resolve an URL with an anonymous token ", function(){
                 expect(Django.url('test_arg', [41])).toBe('/test/arg/41');
             });
 
-            it("should resolve an URL with many anonymous token", function(){
+            it("resolve an URL with many anonymous token", function(){
                 expect(Django.url('test_arg_multi', [41, 'test'])).toBe('/test/arg/41/test');
                 expect(Django.url('test_arg_multi', ['test', 41])).toBe('/test/arg/test/41');
             });
 
-            it("should resolve an URL with a named token", function(){
+            it("resolve an URL with a named token", function(){
                 expect(Django.url('test_named', ['test'])).toBe('/test/named/test');
             });
 
-            it("should resolve an URL with many named token", function(){
+            it("resolve an URL with many named token", function(){
                 expect(Django.url('test_named_multi', [41, 'test'])).toBe('/test/named/41/test');
                 expect(Django.url('test_named_multi', ['test', 41])).toBe('/test/named/test/41');
             });
 
-            it('should throw if number of parameters mismatch', function(){
+            it('throw if number of parameters mismatch', function(){
                 var tooMany = function() {
                     Django.url('test_arg', [1, 2]);
                 };
@@ -144,16 +144,16 @@ describe("Django.js", function(){
         });
 
         describe("from object", function(){
-            it("should resolve an URL without parameter", function(){
+            it("resolve an URL without parameter", function(){
                 expect(Django.url('test_form', {})).toBe('/test/form');
             });
 
-            it("should resolve an URL with named parameters", function(){
+            it("resolve an URL with named parameters", function(){
                 expect(Django.url('test_named', {test: 'value'})).toBe('/test/named/value');
                 expect(Django.url('test_named_multi', {str: 'value', num:41})).toBe('/test/named/value/41');
             });
 
-            it('should throw if there is an anonymous token', function(){
+            it('throw if there is an anonymous token', function(){
                 var anonymous = function() {
                     Django.url('test_arg', {str: 'value'});
                 };
@@ -161,7 +161,7 @@ describe("Django.js", function(){
                 expect(anonymous).toThrow();
             });
 
-            it('should throw if a token is missing', function(){
+            it('throw if a token is missing', function(){
                 var missing = function() {
                     Django.url('test_named_multi', {str: 'value'});
                 };
@@ -169,7 +169,7 @@ describe("Django.js", function(){
                 expect(missing).toThrow();
             });
 
-            it('should not throw for unused key', function(){
+            it("don't throw for unused key", function(){
                 var unused = function() {
                     Django.url('test_named', {test: 'value', num:41});
                 };
@@ -180,21 +180,20 @@ describe("Django.js", function(){
 
         describe('with namespace', function(){
 
-            it("should resolve an URL with a namespace", function(){
-                var namespaceless = function(){
-                    Django.url('fake');
-                };
+            it("resolve an URL with a namespace", function(){
 
                 expect(Django.url('ns1:fake')).toBe('/test/namespace1/fake');
-                expect(namespaceless).toThrow();
 
             });
 
-            it("should not resolve a namespaced url without namespace", function(){
-
+            it("not resolve a namespaced url without namespace", function(){
+                var namespaceless = function(){
+                    Django.url('fake');
+                };
+                expect(namespaceless).toThrow();
             })
 
-            it("should resolve an URL with a nested namespace", function(){
+            it("resolve an URL with a nested namespace", function(){
                 var notLeaf = function(){
                     Django.url('ns2:nested');
                 };
@@ -208,11 +207,11 @@ describe("Django.js", function(){
                 expect(notNested).toThrow();
             });
 
-            it("should resolve an URL with an instance namespace", function(){
+            it("resolve an URL with an instance namespace", function(){
                 expect(Django.url('app1:fake')).toBe('/test/namespace1/fake');
             });
 
-            it("should resolve an URL with a nested instance namespace", function(){
+            it("resolve an URL with a nested instance namespace", function(){
                 expect(Django.url('app2:nested:fake')).toBe('/test/namespace2/nested/fake');
                 expect(Django.url('ns2:appnested:fake')).toBe('/test/namespace2/nested/fake');
                 expect(Django.url('app2:appnested:fake')).toBe('/test/namespace2/nested/fake');
@@ -222,7 +221,7 @@ describe("Django.js", function(){
     });
 
     describe('Resolve static', function(){
-        it("should resolve a static URL", function(){
+        it("resolve a static URL", function(){
             expect(Django.file('my.json')).toBe(DJANGO_INFOS.STATIC_URL + 'my.json');
         });
     });
@@ -230,15 +229,15 @@ describe("Django.js", function(){
     describe('Internationalization', function(){
 
         describe("Automatically include Django provided functions", function(){
-            it('gettext function should be present', function(){
+            it('loads the "gettext" function', function(){
                 expect(gettext).toBeDefined();
             });
 
-            it('ngettext function should be present', function(){
+            it('loads the "ngettext" function', function(){
                 expect(ngettext).toBeDefined();
             });
 
-            it('interpolate function should be present', function(){
+            it('loads the "interpolate" function', function(){
                 expect(interpolate).toBeDefined();
             });
         });
@@ -247,7 +246,7 @@ describe("Django.js", function(){
 
     describe('jQuery Ajax CSRF Helper', function(){
 
-        it("should allow to post Django forms with jQuery Ajax", function(){
+        it("allow to post Django forms with jQuery Ajax", function(){
             var callback = jasmine.createSpy(),
                 $form = $('#test-form'),
                 data = {};
@@ -271,14 +270,14 @@ describe("Django.js", function(){
 
         describe('when django-absolute is missing', function(){
 
-            it("should throw calling absolute()", function(){
+            it("throw calling absolute()", function(){
                 var call = function() {
                     Django.absolute('test_arg', 41);
                 };
                 expect(call).toThrow();
             });
 
-            it("should throw calling site()", function(){
+            it("throw calling site()", function(){
                 var call = function() {
                     Django.site('test_arg', 41);
                 };
@@ -299,11 +298,11 @@ describe("Django.js", function(){
                 delete Django.context.SITE_ROOT;
             });
 
-            it("should get absolute url", function(){
+            it("resolve absolute url", function(){
                 expect(Django.absolute('test_arg', 41)).toBe('http://absolute/test/arg/41');
             });
 
-            it("should get site url", function(){
+            it("resolve site url", function(){
                 expect(Django.site('test_arg', 41)).toBe('http://site/test/arg/41');
             });