+++ CHANGELOG.rst
@@ -7,7 +7,10 @@ Current
 - Rename ``test/js`` into ``js/test`` and reorganize test resources
 - Added namespaced URLs support
 - Handle ``runner_url_args`` and ``runner_url_kwargs`` on ``JsTestCase``
-- renamed ``runner_url*`` into ``url*`` on ``JsTestCase``
+- Renamed ``runner_url*`` into ``url*`` on ``JsTestCase``
+- Added ``JsFileTestCase`` to run tests without live server
+- Renamed ``JasmineMixin`` into ``JasmineSuite``
+- Renamed ``QUnitMixin`` into ``QUnitSuite``
 
 
 0.4 (2012-12-04)